tools/video_downloader/requirements.txt
boto3
yt_dlp
validators
python-dotenv